api/requirements.txt
fastapi
uvicorn
chartjs
psutil
requests
python-dotenv
supabase
sqlalchemy
psycopg2-binary
